[refactor/slices] Toast slice
