+++ src/store/game/toast/tests/hide-toast.test.ts
@@ -1,0 +1,16 @@
+import { expect, it } from "vitest";
+import { ERequestStatus } from "@/shared/enums/requestStatus.enum";
+import { createTestStore } from "./setup/create-test-store";
+
+it("hideToast()", () => {
+  const store = createTestStore();
+  store.setState({
+    toastType: ERequestStatus.ERROR,
+    toastMessage: "Something went wrong",
+  });
+
+  store.getState().hideToast();
+
+  expect(store.getState().toastType).toBeUndefined();
+  expect(store.getState().toastMessage).toBe("");
+});
