+++ src/store/game/toast/tests/setup/create-test-store.ts
@@ -1,0 +1,7 @@
+import { create } from "zustand";
+import { createToastSlice, IToastSlice } from "../..";
+
+export const createTestStore = () =>
+  create<IToastSlice>((...args) => ({
+    ...createToastSlice(...args),
+  }));
