+++ src/store/game/toast/tests/show-toast.test.ts
@@ -1,0 +1,12 @@
+import { expect, it } from "vitest";
+import { ERequestStatus } from "@/shared/enums/requestStatus.enum";
+import { createTestStore } from "./setup/create-test-store";
+
+it("showToast()", () => {
+  const store = createTestStore();
+
+  store.getState().showToast(ERequestStatus.SUCCESS, "Operation successful");
+
+  expect(store.getState().toastType).toBe(ERequestStatus.SUCCESS);
+  expect(store.getState().toastMessage).toBe("Operation successful");
+});
