+++ src/store/game/toast/toast.types.ts
@@ -2,10 +2,11 @@ import { StateCreator } from "zustand";
 import { ERequestStatus } from "@/shared/enums/requestStatus.enum";
 
 export interface IToastSlice {
-  toastType?: ERequestStatus;
   toastMessage: string;
-  showToast: (type: ERequestStatus, message: string) => void;
+  toastType?: ERequestStatus;
+
   hideToast: () => void;
+  showToast: (type: ERequestStatus, message: string) => void;
 }
 
 export type TToastSliceCreator<T extends object> = StateCreator<
